move logging to separate module
